added request_delay_ms config option

+++ src/main/resources/batch_history.sql
@@ -1,5 +1,5 @@
 CREATE TABLE batch_history (
 	dt BIGINT NOT NULL,
-	elapsed INT NOT NULL,
+	elapsed BIGINT NOT NULL,
 	success SMALLINT NOT NULL
 );
